Fix Swoole build deps in installer and resolve Psalm errors
install.sh: add libcurl4-openssl-dev to PHLIX_SWOOLE_BUILD_PKGS — Swoole's
--enable-swoole-curl flag links against libcurl headers (mirrors curl-dev
in phlix-server docker/Dockerfile.base, the build source-of-truth). The
bare-metal Debian/Ubuntu build was missing this, so Swoole failed to
compile. Also documents the configure-flag -> package map.

Psalm (CI errorLevel=1) fixes across src/ (92 -> 0 expected):
- Application.php: correct route/container references from \Phlix\Hub\
  RequestController / RelayController to the real
  \Phlix\Hub\Http\Controllers\ namespace (UndefinedClass).
- Federation repositories: drop the redundant is_array() guard on
  fetchAll() results and keep the @var list<array<string,mixed>> docblock,
  matching the already-clean methods in the same files
  (RedundantCondition / DocblockTypeContradiction).
- Federation DTOs / managers, Http controllers, Hub repositories: narrow
  or cast mixed values from DB rows / request input / settings, add a
  typed class constant, remove redundant int casts, and use strict
  comparisons (MixedAssignment / MixedArgument / RedundantCast /
  RiskyTruthyFalsyComparison / MissingClassConstType / PossiblyFalseArgument).

Behavior-preserving, type-safety-only changes. Verified locally:
php -l clean, PHPCS PSR-12 pass, PHPStan level 6 "No errors". Psalm itself
cannot run on this box (requires PHP >= 8.3.16, host has 8.3.6); CI runs it.

+++ src/Http/Controllers/AuditLogController.php
@@ -46,6 +46,7 @@ final class AuditLogController
         if (isset($params['action']) && is_string($params['action']) && $params['action'] !== '') {
             $filters['action'] = $params['action'];
         }
+        /** @var mixed $successVal */
         $successVal = $params['success'] ?? null;
         if (is_string($successVal) && in_array($successVal, ['0', '1'], true)) {
             $filters['success'] = (bool) (int) $successVal;
@@ -68,8 +69,8 @@ final class AuditLogController
         return (new Response())->json([
             'logs' => $result['entries'],
             'total' => $result['total'],
-            'limit' => (int) ($filters['limit'] ?? 50),
-            'offset' => (int) ($filters['offset'] ?? 0),
+            'limit' => $filters['limit'] ?? 50,
+            'offset' => $filters['offset'] ?? 0,
         ]);
     }
 }
